feat: moving rendering logic out to the ui module
