+++ src/lib.rs
@@ -50,7 +50,7 @@ pub const LOG_LEVEL_ENV_VAR: &str = "AD_LOG";
 
 pub(crate) const VERSION: &str = env!("CARGO_PKG_VERSION");
 pub(crate) const UNNAMED_BUFFER: &str = "[No Name]";
-pub(crate) const MAX_NAME_LEN: usize = 20;
+pub(crate) const MAX_NAME_LEN: usize = 40;
 
 pub(crate) static ORIGINAL_TERMIOS: OnceLock<Termios> = OnceLock::new();
